requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
sqlalchemy
werkzeug
requests
gunicorn